Rename griddly module to griddly_env
