+++ enn_zoo/enn_zoo/griddly_env/images/docs_gdy_path.png
